Refactor dashboard.js: Add settings update route
This commit adds a new route '/settings' to the dashboard.js file. The route is responsible for updating user settings such as file name length, upload password, and hiding user information. When a user makes a POST request to this route, the corresponding user document in the database is updated with the provided settings. If the update is successful, a JSON response with a success message is sent back to the client. In case of any errors, a JSON response with an error message is returned.